- Added glossary.html - Added navigation to Project, Molecule, and Summary pages. - Updated jQuery and plugins to latest and greatest. Table saves works now. - Expanded abbreviated .js files for debugging. Overhead accepted. - Adding attributes to Molecule for storing archive, bmrb and pdb ids. - Adding notes of incomplete validation if no experimental data is present. - Writing a different footer to all html pages. - Started to debug povray rendering. Fails on development machine.

+++ cing/python/cing/PluginCode/molgrap.py
@@ -150,15 +150,17 @@ class Molgrap(NTdict):
         if status:
             nTmessage( "Molgrap.run: rendered/converted entry: %s" % entry_code)
             return True
-
+        # end if
 #        ## Remove temporary files if successful and possible
-        try:
-            os.unlink( pdb_first_file_name )
-            os.unlink( pov_file_name )
-            os.unlink( pov_cor_file_name )
-        except:
-            pass
-
+        if 1: # DEFAULT: True
+            try:
+                os.unlink( pdb_first_file_name )
+                os.unlink( pov_file_name )
+                os.unlink( pov_cor_file_name )
+            except:
+                pass
+            # end try
+        # end if
 
     def _make_molmol_pov_file(self, pdb_file_name, id, backcolor):
         """
